refactor(api): move the dummy-identity helpers to @math3d/api/dev
`dummyIdentity`/`dummyIdToken` forge credentials for allauth's dummy provider,
which only exists on a backend running with `IS_DEPLOYMENT=False`. Exporting
them from the package's main entry put them an autocomplete away from
production code; a subpath entry makes reaching for them deliberate, and
importing one from `@math3d/api` is now a compile error.

The implementation moves to `src/dev/` as well, so the separation is visible in
the tree rather than only in package.json. Follows the shape `@math3d/mock-api`
already uses for `/browser`, `/node` and `/fixtures`.

No production impact either way: with `VITE_ENABLE_DUMMY_AUTH` unset, Vite
folds `ENABLE_DUMMY_AUTH` to false and Rollup drops `DummySignInForm` and both
helpers. The build before and after this commit is byte-identical
(index-C0F5vWcJ.js, 3,459.93 kB).
