+++ packages/app-tests-e2e/src/tests/auth-flows/signin-keeps-scene-edits.test.ts
@@ -1,7 +1,7 @@
 import { test } from "@/fixtures/users";
 import { expect } from "@playwright/test";
 import { SceneBuilder } from "@math3d/mock-api";
-import { dummyIdentity } from "@math3d/api";
+import { dummyIdentity } from "@math3d/api/dev";
 import { makeUserIdentity } from "@/utils/api/auth";
 import AppPage from "@/utils/pages/AppPage";
 import { faker } from "@faker-js/faker/locale/en";
